requirements.txt
django
djangorestframework
django-cors-headers
pillow
python-dotenv
flask
openai
google-generativeai